requirements.txt
streamlit
yfinance
pandas
numpy
pickle
talib
tensorflow
datetime
sklearn.preprocessing
